requirements.txt
haystack-ai==2.0.0b5
streamlit==1.25.0
markdown
python-dotenv
newspaper3k